moved functional test case to integration
